+++ packages/localnet/src/chains/solana/deposit.ts
@@ -2,24 +2,22 @@ import { ethers } from "ethers";
 
 import { NetworkID } from "../../constants";
 import { logger } from "../../logger";
-import { zetachainDepositAndCall } from "../zetachain/zetachainDepositAndCall";
-import { zetachainSwapToCoverGas } from "../zetachain/zetachainSwapToCoverGas";
-import { solanaWithdraw } from "./solanaWithdraw";
-import { solanaWithdrawSPL } from "./solanaWithdrawSPL";
+import { zetachainDeposit } from "../zetachain/deposit";
+import { zetachainSwapToCoverGas } from "../zetachain/swapToCoverGas";
+import { solanaWithdraw } from "./withdraw";
+import { solanaWithdrawSPL } from "./withdrawSPL";
 
-export const solanaDepositAndCall = async ({
-  provider,
+export const solanaDeposit = async ({
   zetachainContracts,
-  args,
+  provider,
   foreignCoins,
+  args,
   deployer,
 }: any) => {
   const chainID = NetworkID.Solana;
   const [sender, , amount, asset] = args;
   try {
-    logger.info("Gateway Deposit and call executed", {
-      chain: NetworkID.Solana,
-    });
+    logger.info("Gateway Deposit executed", { chain: NetworkID.Solana });
     let foreignCoin;
     if (asset === ethers.ZeroAddress) {
       foreignCoin = foreignCoins.find(
@@ -36,14 +34,15 @@ export const solanaDepositAndCall = async ({
       });
       return;
     }
-    await zetachainDepositAndCall({
+
+    await zetachainDeposit({
       args,
       chainID,
       foreignCoins,
-      provider,
       zetachainContracts,
     });
-  } catch (e) {
+  } catch (err) {
+    logger.error(`Error depositing: ${err}`, { chain: NetworkID.ZetaChain });
     const { revertGasFee } = await zetachainSwapToCoverGas({
       amount,
       asset,
