+++ packages/localnet/src/chains/solana/execute.ts
@@ -9,7 +9,7 @@ import { AbiCoder, ethers } from "ethers";
 import { NetworkID } from "../../constants";
 import { logger } from "../../logger";
 import Gateway_IDL from "./idl/gateway.json";
-import { payer, secp256k1KeyPairTSS as tssKeyPair } from "./solanaSetup";
+import { payer, secp256k1KeyPairTSS as tssKeyPair } from "./setup";
 
 export const solanaExecute = async ({
   sender,
