+++ packages/localnet/src/chains/sui/depositAndCall.ts
@@ -2,15 +2,16 @@ import { ethers } from "ethers";
 
 import { NetworkID } from "../../constants";
 import { logger } from "../../logger";
-import { zetachainDeposit } from "../zetachain/zetachainDeposit";
-import { zetachainSwapToCoverGas } from "../zetachain/zetachainSwapToCoverGas";
-import { suiWithdraw } from "./suiWithdraw";
+import { zetachainDepositAndCall } from "../zetachain/depositAndCall";
+import { zetachainSwapToCoverGas } from "../zetachain/swapToCoverGas";
+import { suiWithdraw } from "./withdraw";
 
-export const suiDeposit = async ({
+export const suiDepositAndCall = async ({
   event,
   client,
   deployer,
   foreignCoins,
+  fungibleModuleSigner,
   gatewayObjectId,
   keypair,
   packageId,
@@ -35,20 +36,16 @@ export const suiDeposit = async ({
       : matchingCoin?.asset || ethers.ZeroAddress;
 
   try {
-    logger.info(`Gateway deposit event: ${JSON.stringify(event)}`, {
+    logger.info(`Gateway deposit and call event: ${JSON.stringify(event)}`, {
       chain: chainID,
     });
-    await zetachainDeposit({
-      args: [
-        null,
-        event.receiver,
-        event.amount,
-        event.coin_type === "0x2::sui::SUI"
-          ? ethers.ZeroAddress
-          : event.coin_type,
-      ],
+    const message = ethers.hexlify(new Uint8Array(event.payload));
+    await zetachainDepositAndCall({
+      args: [event.sender, event.receiver, event.amount, asset, message],
       chainID,
       foreignCoins,
+      fungibleModuleSigner,
+      provider,
       zetachainContracts,
     });
   } catch (e) {
@@ -58,6 +55,7 @@ export const suiDeposit = async ({
       chainID,
       deployer,
       foreignCoins,
+      fungibleModuleSigner,
       gasLimit: 200000,
       provider,
       zetachainContracts,
@@ -67,7 +65,6 @@ export const suiDeposit = async ({
       await suiWithdraw({
         amount: revertAmount,
         client: client,
-        coinType: event.coin_type,
         gatewayObjectId: gatewayObjectId,
         keypair: keypair,
         packageId,
